requirements.txt
openai>=1.0.0
rouge_score>=0.1.2
inputimeout>=1.0.0
requests>=2.25.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0
httpx>=0.25.0
pyyaml>=6.0
